feat: ✨ refactor GlobalNotFound component to utilize NotFoundPage for improved structure and clarity
